better duplication check
Fixed the duplication check by removing the original id and generations
numbers (so that the check looked at values rather then placement).

A bug was found and fixed in the duplication check that caused pointers
to be broken - since duplicated objects might have the same values but
different memory space, the pointer system wasn’t persistent. This is
now fixed.