Fixes broken nesting and performance
several performance issues were introduced a few updates ago, and the
code for nesting the << operator was broken

The updates also caused slowdowns on large PDF files (with many pages).

These slowdowns are now resolved.

This update also corrects the reference to the RC4 gem and removes
unnecessary warnings to the console.