Fix: TypeError: can't dup NilClass
In the instance where the catalogs variable does not have a :ProcSet
key, the code generates a new hash and assigns it to that key.  It then
tries to dup the value in that new hash for a key that does not exist.
Other code above uses the inheritance_hash to dup when the data does not
exists and this follows that same flow.